merge saveReturn in GUI_Development

+++ src/Comparators/RentalUserIDComparator.java
@@ -15,7 +15,7 @@ import java.util.Comparator;
 public class RentalUserIDComparator implements Comparator<Rentals>{
     @Override
     public int compare(Rentals r1, Rentals r2) {
-        int c = r1.getUsers_UserID() - r2.getUsers_UserID();
+        int c = (int) r1.getUsers_UserID() -  (int) r2.getUsers_UserID();
         return c;
     }
